+++ src/commands/find_command.cpp
@@ -4,12 +4,15 @@
 
 #include <regex>
 
-const std::regex FIND_REGEX{"[a-zA-Z0-9]*"};
+const std::regex FIND_REGEX{"[- a-zA-Z0-9]*"};
 
-FindCommand::FindCommand() : Command("find", "Find stations by a given search value. Only characters and number are allowed.") {
+FindCommand::FindCommand() : Command("find", "Find stations by name, genre and country.") {
+    m_cli.addOption('n', "name", true, "Search by name");
+    m_cli.addOption('g', "genre", true, "Search by genre");   
+    m_cli.addOption('c', "country", true, "Search by country");   
 }
 
-FindCommand::~FindCommand() { }
+FindCommand::~FindCommand() = default;
 
 void FindCommand::execute(const std::vector<std::string>& args) {
     m_cli.parse(args);
@@ -19,20 +22,24 @@ void FindCommand::execute(const std::vector<std::string>& args) {
         return;
     }
 
-    if (m_cli.getResidualValues().empty()) {
-        LOG(plog::error) << "No search value given. See 'cora find --help' for more information.";
+    if (!std::regex_match(m_cli.getValue('n'), FIND_REGEX)) {
+        LOG(plog::error) << "Invalid search value. See 'cora find --help' for more information.";
         return;
     }
 
-    const std::string value = m_cli.getResidualValues()[1];
-    if (!std::regex_match(value, FIND_REGEX)) {
-        LOG(plog::error) << "Invalid search value. See 'cora find --help' for more information.";
+    if (!std::regex_match(m_cli.getValue('n'), FIND_REGEX)) {
+        LOG(plog::error) << "Invalid value for name. See 'cora find --help' for more information.";
+        return;
+    }
+
+    if (!std::regex_match(m_cli.getValue('g'), FIND_REGEX)) {
+        LOG(plog::error) << "Invalid value for genre. See 'cora find --help' for more information.";
         return;
     }
 
     auto stationsDao = createStationsDao();
     stationsDao->open(m_cli.getValue('f', getDefaultFile()));
-    const auto ids = stationsDao->find(m_cli.getResidualValues()[1]);
+    const auto ids = stationsDao->find(m_cli.getValue('n', ""), m_cli.getValue('g', ""), m_cli.getValue('c', ""));
 
     for (const auto& id : ids) {
         const auto station = stationsDao->findById(id);